lab_1_task_4 3_methods_of_number_processing
Method 1. Find the sum of the non-simple divisors of a number. 
Method 2. Find the number of digits of a number less than 3.
Method 3. Find the number of numbers that are not divisors of the original number, not mutually simple with it and mutually simple with the sum of the prime digits of this number.